requirements.txt
appdirs==1.4.4
attrs==23.2.0
black==23.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
idna==3.6
importlib-metadata==6.7.0
importlib-resources==5.12.0
Jinja2==2.11.2
jsonref==0.2
jsonschema==4.17.3
lark-parser==0.12.0
lxml==5.1.0
mappyfile==0.9.7
#mapview-grids==2.0.20
MarkupSafe==1.1.1
mypy-extensions==0.4.3
OWSLib==0.29.3
packaging==23.2
pathspec==0.11.2
pep8==1.7.1
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
pycodestyle==2.10.0
pyodbc==4.0.30
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
pyqt5-tools==5.15.9.3.3
PyQt5_sip==12.15.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.15.0
pytz==2024.1
PyYAML==6.0.1
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
regex==2021.4.4
requests==2.31.0
six==1.16.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.7.1
urllib3==1.26.2
xmltodict==0.13.0
zipp==3.15.0
